split out manifest loading
